refactor(cli): migrate onboard-session.js to TypeScript
Convert bin/lib/onboard-session.js (440 lines) to src/lib/onboard-session.ts
with typed interfaces for Session, StepState, SessionFailure, LockResult,
SessionUpdates, and related types.

Full migration — all functions including fs I/O (session persistence,
file-based locking) and pure helpers (redaction, validation, normalization).

Co-locates tests. Cache-clearing pattern updated for dist/ path.

615 CLI tests pass. Coverage ratchet passes.

Relates to #924 (shell consolidation).
